touchcore/rest/serializers.py
from rest_framework import serializers
from touchcore.rest.models import Employee, Company

class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = '__all__'

class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = ['id', 'name', ]
