requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
cloudinary==1.19.1
dj-database-url==0.5.0
Django==3.0.1
django-cloudinary-storage==0.2.3
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
mock==3.0.5
Pillow==6.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
